edgar/pyproject.toml
[project]
name = "edgar"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "azure-identity>=1.20.0",
    "azure-storage-blob>=12.24.1",
    "edgartools>=3.12.1",
    "loguru>=0.7.3",
    "pandas>=2.2.3",
    "pydantic>=2.10.6",
    "pydantic-settings>=2.8.1",
    "python-dotenv>=1.0.1",
    "tqdm>=4.67.1",
]
